automation/requirements.txt
pandas
reportlab
matplotlib
python-dotenv
schedule
openpyxl
playwright
pyodbc
